requirements.txt
autopep8==1.6.0
bidict==0.21.4
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
dnspython==1.16.0
eventlet==0.30.2
Flask==2.0.2
Flask-Mail==0.9.1
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
fpdf==1.7.2
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.21.4
psycopg2-binary==2.9.2
pycodestyle==2.8.0
PyJWT==2.3.0
python-dotenv==0.19.2
python-engineio==4.3.0
python-socketio==5.5.0
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.2
simple-websocket==0.4.0
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.27
threadpoolctl==3.0.0
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
wsproto==1.0.0
